+++ base/compilerimg.jl
@@ -1,0 +1,4 @@
+# This file is a part of Julia. License is MIT: https://julialang.org/license
+
+baremodule Base; end
+Core.include(Base, "Base_compiler.jl")
