+++ base/essentials.jl
@@ -183,11 +183,7 @@ end
 _nameof(m::Module) = ccall(:jl_module_name, Ref{Symbol}, (Any,), m)
 
 function _is_internal(__module__)
-    if ccall(:jl_base_relative_to, Any, (Any,), __module__)::Module === Core.Compiler ||
-       _nameof(__module__) === :Base
-        return true
-    end
-    return false
+    return true
 end
 
 # can be used in place of `@assume_effects :total` (supposed to be used for bootstrapping)
